moved human class to player.py
