+++ src/__init__.py
@@ -4,7 +4,6 @@ __all__ = ["Game", "Display", "Board", "Human", "Bot", "Player", "config"]
 from .game import Game
 from .display import Display
 from .board import Board
-from .human import Human
 from .bot import Bot
-from ._player import Player
+from .player import Player
 from .config import *
